tpu-examples/single-host-inference/jax/requirements.txt
requests
numpy
tensorflow
pillow
tensorflow-serving-api
transformers
diffusers
flask